lag.zig: use simd + mixin methods
As you can likely guess, i'm tw0st3p on twitch.  This is just a fun
patch to see if i could make lag work with simd and also use a mixin
struct to get rid of some repitition.  I think its pretty neat.  But
ofcourse, take it or leave it.

Not sure but I imagine it would be more performant even with the extra
padding field in Vec3 due to using simd instructions.